Extract integer validator into dedicated Trait
In order to ease Unit Test redability
