+++ tests/units/Stub/UI/Resolver/Validator/Implementation/AbstractValidatorStub.php
@@ -1,0 +1,30 @@
+<?php
+declare(strict_types = 1);
+
+namespace Tests\Units\Imedia\Ammit\Stub\UI\Resolver\Validator\Implementation;
+
+use Imedia\Ammit\UI\Resolver\Exception\UIValidationException;
+use Imedia\Ammit\UI\Resolver\UIValidationEngine;
+use Imedia\Ammit\UI\Resolver\Validator\UIValidatorInterface;
+
+/**
+ * @author Guillaume MOREL <g.morel@imediafrance.fr>
+ */
+abstract class AbstractValidatorStub implements UIValidatorInterface
+{
+    /** @var UIValidationEngine */
+    protected $validationEngine;
+
+    public function __construct(UIValidationEngine $validationEngine)
+    {
+        $this->validationEngine = $validationEngine;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function createUIValidationException(string $message, string $propertyPath = null): UIValidationException
+    {
+        return UIValidationException::fromRaw($message, $propertyPath);
+    }
+}
